requirements.txt
fastapi
uvicorn
psycopg2
pandas
langgraph
sqlalchemy
